test compare_spec
