+++ pyproject.toml
@@ -26,6 +26,7 @@ dependencies = [
     "dvc-s3>=3.2.0",
     "mpcontribs-client>=5.10.2",
     "pydantic>=2.10.6",
+    "deepdiff>=8.5.0",
 ]
 
 [dependency-groups]
